requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.0.1
httpx==0.27.2
openai==1.52.0
fal-client==0.5.9
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
